feat(B-143): rename app review files
